ADD: refactor ReplaceObjectIds
