+++ westack/datasource/datasource.go
@@ -5,8 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	"regexp"
-	"strconv"
 	"time"
 
 	"github.com/go-redis/redis/v8"
@@ -14,7 +12,6 @@ import (
 	"github.com/spf13/viper"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/bsoncodec"
-	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/event"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
@@ -505,238 +502,4 @@ func New(dsKey string, dsViper *viper.Viper, parentContext context.Context) *Dat
 		ctxCancelFn: ctxCancelFn,
 	}
 	return ds
-}
-
-func ReplaceObjectIds(data interface{}) interface{} {
-
-	if data == nil {
-		return nil
-	}
-
-	var finalData wst.M
-	switch data.(type) {
-	case int, int32, int64, float32, float64, bool, primitive.ObjectID, *primitive.ObjectID, time.Time, primitive.DateTime:
-		return data
-	case string:
-		var newValue interface{}
-		var err error
-		dataSt := data.(string)
-		if wst.RegexpIdEntire.MatchString(dataSt) {
-			newValue, err = primitive.ObjectIDFromHex(dataSt)
-		} else if wst.IsAnyDate(dataSt) {
-			newValue, err = wst.ParseDate(dataSt)
-		}
-		if err != nil {
-			log.Println("WARNING: ", err)
-		}
-		if newValue != nil {
-			return newValue
-		} else {
-			return data
-		}
-	case wst.Where:
-		finalData = wst.M{}
-		for key, value := range data.(wst.Where) {
-			finalData[key] = value
-		}
-		break
-	case *wst.Where:
-		finalData = wst.M{}
-		for key, value := range *data.(*wst.Where) {
-			finalData[key] = value
-		}
-		break
-	case map[string]interface{}:
-		finalData = wst.M{}
-		for key, value := range data.(map[string]interface{}) {
-			finalData[key] = value
-		}
-		break
-	case *map[string]interface{}:
-		finalData = wst.M{}
-		for key, value := range *data.(*map[string]interface{}) {
-			finalData[key] = value
-		}
-		break
-	case wst.M:
-		finalData = data.(wst.M)
-		break
-	case *wst.M:
-		finalData = *data.(*wst.M)
-		break
-	default:
-		log.Println(fmt.Sprintf("WARNING: Invalid input for ReplaceObjectIds() <- %s", data))
-		return data
-	}
-	for key, value := range finalData {
-		if value == nil {
-			continue
-		}
-		var err error
-		var newValue interface{}
-		if key == "$eq" || key == "$ne" || key == "$gt" || key == "$gte" || key == "$lt" || key == "$lte" {
-			if value == "$now" {
-				newValue = time.Now()
-			} else if value == "$today" {
-				newValue = time.Now().Truncate(24 * time.Hour)
-			} else if value == "$yesterday" {
-				newValue = time.Now().Truncate(24 * time.Hour).Add(-24 * time.Hour)
-			} else if value == "$tomorrow" {
-				newValue = time.Now().Truncate(24 * time.Hour).Add(24 * time.Hour)
-			} else {
-				switch value.(type) {
-				case string:
-					if r := regexp.MustCompile(`^\$(\d+)dago$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(-time.Duration(atoi) * 24 * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)wago$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(-time.Duration(atoi) * 7 * 24 * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)mago$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(-time.Duration(atoi) * 30 * 24 * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)yago$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(-time.Duration(atoi) * 365 * 24 * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)Sago$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(-time.Duration(atoi) * time.Second)
-					} else if r := regexp.MustCompile(`^\$(\d+)Mago$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(-time.Duration(atoi) * time.Minute)
-					} else if r := regexp.MustCompile(`^\$(\d+)Hago$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(-time.Duration(atoi) * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)dfromnow$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(time.Duration(atoi) * 24 * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)wfromnow$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(time.Duration(atoi) * 7 * 24 * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)mfromnow$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(time.Duration(atoi) * 30 * 24 * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)yfromnow$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(time.Duration(atoi) * 365 * 24 * time.Hour)
-					} else if r := regexp.MustCompile(`^\$(\d+)Sfromnow$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(time.Duration(atoi) * time.Second)
-					} else if r := regexp.MustCompile(`^\$(\d+)Mfromnow$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(time.Duration(atoi) * time.Minute)
-					} else if r := regexp.MustCompile(`^\$(\d+)Hfromnow$`).FindStringSubmatch(value.(string)); len(r) > 1 {
-						atoi, err := strconv.Atoi(r[1])
-						if err != nil {
-							panic(err)
-						}
-						newValue = time.Now().Add(time.Duration(atoi) * time.Hour)
-					}
-				}
-			}
-		}
-		if newValue == nil {
-			switch value.(type) {
-			case string, wst.Where, *wst.Where, wst.M, *wst.M, int, int32, int64, float32, float64, bool, primitive.ObjectID, *primitive.ObjectID, time.Time, primitive.DateTime:
-				newValue = ReplaceObjectIds(value)
-				break
-			default:
-				asMap, asMapOk := value.(wst.M)
-				if asMapOk {
-					newValue = ReplaceObjectIds(asMap)
-				} else {
-					asList, asListOk := value.([]interface{})
-					if asListOk {
-						for i, asListItem := range asList {
-							asList[i] = ReplaceObjectIds(asListItem)
-						}
-					} else {
-						_, asStringListOk := value.([]string)
-						if !asStringListOk {
-							asMap, asMapOk := value.(map[string]interface{})
-							if asMapOk {
-								newValue = ReplaceObjectIds(asMap)
-							} else {
-								asList, asMListOk := value.([]wst.M)
-								if asMListOk {
-									for i, asListItem := range asList {
-										asList[i] = ReplaceObjectIds(asListItem).(wst.M)
-									}
-								} else {
-									log.Println(fmt.Sprintf("WARNING: What to do with %v (%s)?", value, value))
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-		if err == nil && newValue != nil {
-			switch data.(type) {
-			case wst.Where:
-				data.(wst.Where)[key] = newValue
-				break
-			case *wst.Where:
-				(*data.(*wst.Where))[key] = newValue
-				break
-			case wst.M:
-				data.(wst.M)[key] = newValue
-				break
-			case *wst.M:
-				(*data.(*wst.M))[key] = newValue
-				break
-			case map[string]interface{}:
-				data.(map[string]interface{})[key] = newValue
-				break
-			case *map[string]interface{}:
-				(*data.(*map[string]interface{}))[key] = newValue
-				break
-			default:
-				log.Println(fmt.Sprintf("WARNING: invalid input ReplaceObjectIds() <- %s", data))
-				break
-			}
-		} else if err != nil {
-			log.Println("WARNING: ", err)
-		}
-	}
-	return data
 }
